requirements.txt
opencv-python==4.13.0.90
pillow==12.1.0
numpy==2.4.1
